+++ platforms/common/application/configurations/dropdown-edit.js
@@ -5,11 +5,39 @@ var $             = require('elements'),
     request       = require('agent'),
 
     modal         = require('../ui').modal,
+    guid          = require('mout/random/guid'),
     trim          = require('mout/string/trim'),
 
     getAjaxSuffix = require('../utils/get-ajax-suffix'),
     parseAjaxURI  = require('../utils/get-ajax-url').parse,
-    getAjaxURL    = require('../utils/get-ajax-url').global;
+    getAjaxURL    = require('../utils/get-ajax-url').global,
+
+    History       = require('../utils/history'),
+    getParam      = require('mout/queryString/getParam'),
+    setParam      = require('mout/queryString/setParam');
+
+
+var refreshWordpressLinks = function (title, value) {
+    if (GANTRY_PLATFORM == 'wordpress') {
+        // refresh URIs with new configuration name
+        var replace = title.replace(/[^a-z\d_-\s]/i, '_').toLowerCase(),
+            find = $('[href*="/' + value + '/"]'),
+            currentURI = History.getPageUrl(),
+            currentView = getParam(currentURI, 'view');
+
+        if (find) {
+            find.forEach(function(lnk){
+                lnk = $(lnk);
+                var href = lnk.href().replace('/' + value + '/', '/' + replace + '/');
+                lnk.href(href);
+            });
+        }
+
+        currentView = currentView.replace('/' + value + '/', '/' + replace + '/');
+        currentURI = setParam(currentURI, 'view', currentView);
+        History.replaceState({ uuid: guid(), doNothing: true }, window.document.title, currentURI);
+    }
+}
 
 ready(function() {
     var body = $('body');
@@ -64,6 +92,9 @@ ready(function() {
                         selectized.style('display', 'inline-block');
                         editable.style('display', 'none')
                     }
+
+                    // fix Wordpress non unique IDs by refreshing all hrefs
+                    refreshWordpressLinks(title, value);
 
                     element.removeClass('disabled');
                     element.removeClass('fa-spin-fast fa-spinner').addClass('fa-pencil');