+++ platforms/common/js/main.js
@@ -370,11 +370,39 @@ var $             = require('elements'),
     request       = require('agent'),
 
     modal         = require('../ui').modal,
+    guid          = require('mout/random/guid'),
     trim          = require('mout/string/trim'),
 
     getAjaxSuffix = require('../utils/get-ajax-suffix'),
     parseAjaxURI  = require('../utils/get-ajax-url').parse,
-    getAjaxURL    = require('../utils/get-ajax-url').global;
+    getAjaxURL    = require('../utils/get-ajax-url').global,
+
+    History       = require('../utils/history'),
+    getParam      = require('mout/queryString/getParam'),
+    setParam      = require('mout/queryString/setParam');
+
+
+var refreshWordpressLinks = function (title, value) {
+    if (GANTRY_PLATFORM == 'wordpress') {
+        // refresh URIs with new configuration name
+        var replace = title.replace(/[^a-z\d_-\s]/i, '_').toLowerCase(),
+            find = $('[href*="/' + value + '/"]'),
+            currentURI = History.getPageUrl(),
+            currentView = getParam(currentURI, 'view');
+
+        if (find) {
+            find.forEach(function(lnk){
+                lnk = $(lnk);
+                var href = lnk.href().replace('/' + value + '/', '/' + replace + '/');
+                lnk.href(href);
+            });
+        }
+
+        currentView = currentView.replace('/' + value + '/', '/' + replace + '/');
+        currentURI = setParam(currentURI, 'view', currentView);
+        History.replaceState({ uuid: guid(), doNothing: true }, window.document.title, currentURI);
+    }
+}
 
 ready(function() {
     var body = $('body');
@@ -430,6 +458,9 @@ ready(function() {
                         editable.style('display', 'none')
                     }
 
+                    // fix Wordpress non unique IDs by refreshing all hrefs
+                    refreshWordpressLinks(title, value);
+
                     element.removeClass('disabled');
                     element.removeClass('fa-spin-fast fa-spinner').addClass('fa-pencil');
                 });
@@ -446,7 +477,7 @@ ready(function() {
 
 module.exports = {};
 
-},{"../ui":46,"../utils/get-ajax-suffix":60,"../utils/get-ajax-url":61,"agent":67,"elements":98,"elements/domready":96,"mout/string/trim":233}],5:[function(require,module,exports){
+},{"../ui":46,"../utils/get-ajax-suffix":60,"../utils/get-ajax-url":61,"../utils/history":64,"agent":67,"elements":98,"elements/domready":96,"mout/queryString/getParam":209,"mout/queryString/setParam":211,"mout/random/guid":213,"mout/string/trim":233}],5:[function(require,module,exports){
 "use strict";
 
 var $             = require('elements'),
@@ -12816,7 +12847,6 @@ domready(function() {
         if (currentNonce !== GANTRY_AJAX_NONCE) {
             currentURI = setParam(currentURI, '_wpnonce', GANTRY_AJAX_NONCE);
             History.replaceState({ uuid: guid(), doNothing: true }, window.document.title, currentURI);
-            window.HHH = History;
         }
     }