+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 1. [Common](#common)
     3. [](#bugfix)
         - Fixed issue in admin where overrides for Enabled toggle wouldn't be showing checked, causing the value to reset to Base Outline
+3. [WordPress](#wordpress)
+    3. [](#bugfix)
+        - Renaming of Outlines from navigation bar will now properly refresh all links with the new value (#912)
+
 # 5.1.6
 ## 14/10/2015